Add a general way of building repositories from reducers
The only mechanism of inputing data is by persisting events. All other
data types are essentially views of the data, computed from the events.
This means that repositories showing the current state of these computed
data types can be built from an event repository and a reducer that
transforms a list of events into a state.